To update local version from ulex-opensource (#47)
* Set theme jekyll-theme-midnight

* Added back references 8 and 9

* Footnote backref test 2

* Fixed rest of back refs

* Fixed footnotes in release notes

* Custom layout (#3)

* Added Ulex logo

* Added midnight layout

* Added logo to layout

* Removed layout header

* Update to procedural rules for version 1.2

This edit aims to make the procedural rules more user-friendly as well as to add a new rule for appeals.  The latter edit justifies releasing this as Ulex v. 1.2.  There will be other edits in that version; this is just the first of several.

* Removal of governmental immunity for v. 1.2

This change nixes rules that might otherwise immunize government actors or, per the operation of Ulex rule 3.2, parties representing "the closest functionally equivalent institution [or] office . . . ."

* Typo fix and two updates for v. 1.2

* Wrongful death and survival act

This adds to v. 1.2 a wrongful death and survival act--necessary to ensure that tort claims do not expire with their victims.

* Update README.md

* Update README.md

* Revision to 1.2.2. to avoid hung panels

* Changing "may" to "must" ...

... in Rule 1.2.2. to clarify that judges have only two options:  choose a proposed remedy else dismiss it.  The dismissal may be by direct vote, as by a judge that finds both proposed remedies (mutual requests to torture the counter-party, for instance) illegal, or by inaction, as by judges that fail to reach a majority concerning either proposed remedy.

* Revert to light theme (#7)

* Reverted to slate theme

* Set theme jekyll-theme-slate

* Added Ulex logo

* Fixed logo alignment

* Added v1.2

* Switched top link to v 1.2

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Create CNAME

* Update CNAME

* Update README.md

* Corrects links in Rules 2.6.3 and 2.7.3

Trying to use web.archive.org versions when possible, as live links will more likely rot.

* Update README.md

The edit fixes a broken link.  The new link still does not provide the cited rules, but does at least provide context that might prove useful in continuing the search.

Co-authored-by: David McFadzean <davidmc@gmail.com>
Co-authored-by: joestartupsocieties <50885181+joestartupsocieties@users.noreply.github.com>